+++ router.h
@@ -20,16 +20,16 @@ public:
     const uint32_t router_id;
 
     // map from inport to its predicate map
-    std::map< uint32_t, std::map<Json::Value, PredicateNode*>* > predicate_map;
+    std::map< uint64_t, std::map<Json::Value, PredicateNode*>* > predicate_map;
 
     // map from inport to the packet header space already dealt with on this in_port
-    std::map< uint32_t, bdd > dealt_bdd_map;
+    std::map< uint64_t, bdd > dealt_bdd_map;
 
     // map from inport to its ap represented in bool vector;
-    std::map< uint32_t, std::map<Json::Value, APNodeV*>* > ap_vec_map;
+    std::map< uint64_t, std::map<Json::Value, APNodeV*>* > ap_vec_map;
 
     // map from inport to its ap represented in bitset;
-    std::map< uint32_t, std::map<Json::Value, APNodeB*>* > ap_bset_map;
+    std::map< uint64_t, std::map<Json::Value, APNodeB*>* > ap_bset_map;
 
     // constructor
     Router(uint32_t id);
@@ -51,8 +51,5 @@ public:
 
     void print_router_apn_map();
 
-//    void propagate(std::vector< bool > packet_header, std::list< uint32_t > passed_port, uint32_t dst_port,
-//                   uint32_t inport, std::map< uint32_t, std::vector<uint32_t>* >* topology,
-//                   std::map< uint32_t, uint32_t >* inport_to_router, std::map< uint32_t, Router* >* id_to_router);
 };
 #endif //AP_VERIFIER_ROUTER_H
